make.sh: fix warning when deleting .config at blank build

+++ make.sh
@@ -224,7 +224,9 @@ build_minibase()
 config_init()
 {
 	if [ $DISTRO_DEFCONFIG ] && [ -e $CONFIGS_DIR/$DISTRO_DEFCONFIG ];then
-		rm $DISTRO_CONFIG
+		if [ -e $DISTRO_CONFIG ];then
+			rm $DISTRO_CONFIG
+		fi
 		for line in $(cat $CONFIGS_DIR/$DISTRO_DEFCONFIG)
 		do
 			if [ $line ] && [ -e $CONFIGS_DIR/$line ];then
